Use old_sets.bzl where sets are needed so that we do not have to pin (#2208)
skylib to 0.8.0 or below. This allows interoperability with the federation.
Federation users will end up bringing in bazel-skylib at 0.9.0, because
they (by default) pin the version before rules_go.

This was tested along with an update to bazel-skylib 0.9.0. This was to
prove that the old_sets change will continue to do the right thing with
the next skylib release. We could combine that into this release, but
it is not strictly necessary.

Updates #2157