go/tools/bazel_testing: support location args without external prefix (#2287)
Previously, Bazel arguments resolved with $(location) or $(locations)
always included an "external/workspace/" prefix. We used this prefix
to map arguments to runfiles. Upcoming versions of Bazel won't add
this prefix to files in the main workspace.

Witht this change, bazel_testing will read the workspace name from the
WORKSPACE file from the main workspace (which must be among the
arguments). This makes sure that go_bazel_test targets in the main
workspace will continue to work.

Updates #2285