go: disable race and msan modes when pure mode is set (#2506)
In go_transition, race and msan settings will be disabled if pure mode
is explicitly enabled. If race or msan are explicitly enabled together
with pure mode, an error will be reported, but this isn't fatal.

In get_mode, --features=race and --features=msan will be ignored when
pure mode is enabled. If race or msan are explicitly enabled together
with pure mode, a fatal error will be reported.

Fixes #2499