proto: wrap pre-generated WKT libraries with go_proto_library (#2511)
Several of the well-known types have pre-generated code checked into
to both github.com/golang/protobuf (APIv1) and
google.golang.org/protobuf (APIv2). The APIv1 packages are just
wrappers for the APIv2 packages. We're using the APIv1 compiler, but
it won't generate these wrappers for us automatically. If we use it,
and a program also imports the APIv2 packages, we'll end up with a
registration conflict.

Instead, we'll use the pre-generated APIv1 code, but we'll wrap it up
in a go_proto_library so all the usual aspects and providers will
continue working.

Fixes #2501