link: find package conflict errors during analysis, report in GoLink (#2553)
In general, packages should be compiled with the same dependencies
they're linked against. We haven't been too strict about this in the
past, and there are two common situations where this might not happen:
1) go_proto_library and go_library targets are linked for the same
package, 2) an external test imports a library that imports the
package being tested.

Analysis will now detect both problems and will pass a descriptive
message to the GoLink action. Making this a hard error would be an
incompatible change, so for now, it will be reported if the
incompatible_package_conflict_is_error setting is enabled (#1374) or
if the linker fails.

Fixes #1877
Updates #1374
Updates #2535