[bazeltesting] Refactor to SpliceOSArgs utility function (#2887)
Fixes #2886

This PR replaces #2882 (that one was done using github editor). Based on feedback, created this PR.

Note, this function cannot easily be put into the bazeltesting package because it already has a TestMain. As such, I put it in the bazel package where a number of utility functions used by bazeltesting already exist.