Make go_test() with embedsrcs work (#2936)
Even though there already is testing coverage for go_test() with
embedsrcs using generated files and such, I can't seem to get one of the
most trivial cases to work. The file to be embedded doesn't get placed
in the sandbox during GoCompilePkg.

This change addresses this by explicitly copying the embedsrcs from the
internal source to the external, which seems to make things work
properly.