go_path: Support go:embed (#3163)
* go_path: Support go:embed

The `go_path` rule does not currently copy embedded sources.
This means that if we use `go:embed` in any file and use `go_path` to
build a GOPATH from that, the resulting code will not compile.
Example of an error from one of our internal systems:

    scaffold.go:9:13: pattern template/*: no matching files found

To fix this, the go_path rule needs to copy the embedsrcs over from the
archive and add them to the manifest for the go_path builder to copy.

* Handle embedsrcs inside subdirectories

The previous implementation of this did not handle a `go:embed` for a
subdirectory that was not a Bazel package. For example,

    mypkg/
    |- BUILD.bazel
    |- foo.go   // contains "go:embed template/foo"
    |- template/
        |- foo

With this, the BUILD.bazel for mypkg will hold
`embedsrcs = ["template/foo"]`.

However, if we just append `f.basename` to the output directory path,
this will generate `mypkg/foo`, not `mypkg/template/foo`.

To fix this, this commit is tracking the directory in which the last
source file was found,
and use a path to the embedded file relative to that directory.