Refactor away references to @io_bazel_rules_go (#3185)
* Use repo-relative labels in load statements

* Introduce a constant for the Go toolchain requirement

Gets rid of a common source of repo-absolute labels. Repo-relative
labels for toolchains aren't supported with Bazel 4. Now that
config_common.toolchain_type exists, the string label may also be
replaced with a more complex Starlark type in the future.

* Use Label(...) in macros

This resolves a label literal relative to the call site of the Label
constructor.

* Use repo-relative labels in rule attribute defaults