link.bzl: ignore duplicate dep on coverdata (#3032)
* link.bzl: ignore duplicate dep on coverdata

Importing the coverdata package from user code caused the link to fail:

  link: package conflict error: github.com/bazelbuild/rules_go/go/tools/coverdata: multiple copies of package passed to linker:
          @io_bazel_rules_go//go/tools/coverdata:coverdata
          @io_bazel_rules_go//go/tools/coverdata:coverdata

Two different archives were indeed being provided as the object files for that package:

  -arc '@io_bazel_rules_go//go/tools/coverdata:coverdata=github.com/bazelbuild/rules_go/go/tools/coverdata=bazel-out/darwin_arm64-fastbuild/bin/external/io_bazel_rules_go/go/tools/coverdata/coverdata.a'
  -arc '@io_bazel_rules_go//go/tools/coverdata:coverdata=github.com/bazelbuild/rules_go/go/tools/coverdata=bazel-out/darwin_arm64-fastbuild-ST-8f6eb6207b9d/bin/external/io_bazel_rules_go/go/tools/coverdata/coverdata.a'

This is due to a combination of two things:

* coverdata is defined as a `go_tool_library` because it is dependended upon by the toolchain.
  Declaring it as a `go_library` would cause a circular dependency.

* nogo support is implemented with a transition.
  This avoids a circular dependency that requires nogo to build nogo.

Either due to being a `go_tool_library` or due to being a dependent of
the toolchain, coverdata is included on both sides of the transition
when also imported by user code. Since the built package in each
configuration corresponds to different object files, this leads to
the duplication.

In this case, it does not seem important to select a specific
coverdata package (instrumented or not), and it does not appear to be
straightforward to determine which is which. Take the easy way out and
have the policy be: first one in the depset wins.

Since it's not clear if this occurs with other packages (e.g. nogo analyses),
log a warning when a duplicate archive is being ignored.

Fixes #3017

* Update go/private/actions/link.bzl

* Update BUILD.bazel

* Fix buildifier warning

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>
Co-authored-by: Zhongpeng Lin <zplin@uber.com>