Adding first example (#3317)
* Adding first example

This PR includes the source code for the tutorial that is hosted
here: https://bazel-contrib.github.io/SIG-rules-authors/go-tutorial.html

The code for the above html is on the gh-pages branch of the
SIG-rules-authors repository:
https://github.com/bazel-contrib/SIG-rules-authors/tree/gh-pages

This tutorial source code includes:

- basic go rules support
- a basic cobra CLI application
- unit test
- Gazelle support

After this PR is finalized we I will update the tutorial markdown.

We are also adding CI/CD for the example.  This PR modifies the
presubmit.yml file.  This project also uses the words file that
requires the installation of wamerican.

* Updates for PR review

- renamed the tutorial to basic-gazelle
- updated roll_dice to use instead of the word generator

* Updating presubmit for rename

* More Cleanup

- removing the word example code
- creating testing code for roll dice

* Added comment to file per code review

* Updated go mod

- removed dep that we do not need
- changed path

* removing dep

* Adding klog as the dependency

- Imported klog
- added klog to different go files
- used bazel to update go.mod
- used gazelle to updated the BUILD files

* Adding macos test

Adding testing the examples with the macos platform.

* Adding windows test