Make C++ toolchain optional (#3390)
* Make C++ toolchain optional

Currently, if you transition a go_binary, you must have a C++ toolchain
for it even if it's building in pure mode without cgo.

By making this toolchain optional, we push the error from a
toolchain-resolution-time error to an analysis-time error only if a
toolchain is both _needed_ and not found.

* Add tests

* Add target that actually uses cgo

* Switch to use bazel_features

Also switch from @bazel_tools//tools/cpp:current_cc_toolchain to
@bazel_tools//tools/cpp:optional_current_cc_toolchain because otherwise
we still get a hard error if the toolchain can't be found.

* Review comments