feat(mode): add `purego` tag when `pure` (#3901)
The community has agreed[1] that "pure" builds of Go code should
use the `purego` build tag.

This change adds that de-facto tag.

The minimum version of Gazelle is bumped to `0.36.0` which includes a patch[2]
for ignoring the `purego` tag so that the source flies are passed to the
compiler for conditional compilation decisions.

[1]: https://github.com/golang/go/issues/23172
[2]: https://github.com/bazelbuild/bazel-gazelle/pull/1767