Remove unneeded packages from stdlib (#4093)
**What type of PR is this?**
perf improvement

**What does this PR do? Why is it needed?**
This cuts down the time to build the stdlib from 8.1 to 7.4s in my
testing and it also reduces the size of the `sdk.libs` depset and the
resulting linkcfg files.

I wasn't sure if it's ok to exclude all of `cmd/`; I suspect it should
be ok since the toolchain has prebuilt tools anyway, in which case we
can simplify the glob pattern.

**Which issues(s) does this PR fix?**

Fixes #

**Other notes for review**