Support .syso files in the `srcs` lists for `go_library` and `go_test` rules. (#4151)
**What type of PR is this?**
Feature

**What does this PR do? Why is it needed?**

Essentially, `go build ./...` and `go test ./...` allow having `.syso`
files in the packages and it includes them in the packages it creates so
that functions defined therein can be called from Go code.

Bazel didn't allow them to be among the `srcs` in these rules until this
PR.

**Which issues(s) does this PR fix?**

Fixes #4148

**Other notes for review**

I also added a bunch of tests and ran them on both linux (amd64) and Mac
OS (arm64).

It's quite possible that I need to add some more things to get these
tests to work on other platforms which are in the CI system.