Merge GoLibrary and GoSource providers (#4030)
**What type of PR is this?**
Code simplification

**What does this PR do? Why is it needed?**
These providers are 1:1, and we always create a library followed by
`source_from_library`. Merging them allows to remove the `library` field
from `GoSource` as well as the the `resolve` field, since we don't need
to smuggle the closure between the providers anymore.

If we do this, we will also need to fix up Gazelle. We can either have a
patch in this repo so CI passes and then followup with the Gazelle fix,
or somehow shim the API until we release new versions in lockstep. Open
to ideas how to best do this.

**Which issues(s) does this PR fix?**
One step toward https://github.com/bazelbuild/rules_go/issues/2578

Fixes #

**Other notes for review**